# src/main/resources/openapi/bankcards-api.yaml
openapi: 3.0.0
info:
  title: API Сервиса управления картами
  version: 1.0.0
  description: API для CRUD операций с картами и переводов

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    CardStatus:
      type: string
      enum:
        - ACTIVE
        - BLOCKED
        - EXPIRED

    RoleType:
      type: string
      enum:
        - ROLE_USER
        - ROLE_ADMIN

    CardDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        maskedCardNumber:
          type: string
          pattern: '^\*\*\*\* \*\*\*\* \*\*\*\* \d{4}$'
          example: "**** **** **** 1234"
        cardholderName:
          type: string
          example: "John Doe"
        expiryDate:
          type: string
          format: date
          example: "2025-12-31"
        status:
          $ref: '#/components/schemas/CardStatus'
        balance:
          type: number
          format: double
          example: 1000.00
        userId:
          type: string
          format: uuid

    CreateCardRequest:
      type: object
      required:
        - cardholderName
        - expiryDate
        - initialBalance
        - userId
      properties:
        cardholderName:
          type: string
          minLength: 2
          maxLength: 100
        expiryDate:
          type: string
          format: date
        initialBalance:
          type: number
          format: double
          minimum: 0
        userId:
          type: string
          format: uuid

    UpdateCardStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          $ref: '#/components/schemas/CardStatus'

    TransferRequest:
      type: object
      required:
        - fromCardId
        - toCardId
        - amount
      properties:
        fromCardId:
          type: string
          format: uuid
        toCardId:
          type: string
          format: uuid
        amount:
          type: number
          format: double
          minimum: 0.01
        description:
          type: string
          maxLength: 200

    TransactionDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        fromCardId:
          type: string
          format: uuid
        fromCardMask:
          type: string
          example: "**** **** **** 1234"
        toCardId:
          type: string
          format: uuid
        toCardMask:
          type: string
          example: "**** **** **** 5678"
        amount:
          type: number
          format: double
        transactionDate:
          type: string
          format: date
        description:
          type: string

    CardFilterRequest:
      type: object
      properties:
        maskedCardNumber:
          type: string
        status:
          $ref: '#/components/schemas/CardStatus'
        minBalance:
          type: number
          format: double
        maxBalance:
          type: number
          format: double

    UserDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstname:
          type: string
        lastname:
          type: string
        middlename:
          type: string
        enabled:
          type: boolean
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleType'

    UpdateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        firstname:
          type: string
        lastname:
          type: string
        middlename:
          type: string
        enabled:
          type: boolean
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleType'

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string

    GetMyCards200Response:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/CardDTO'
        totalElements:
          type: integer
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean

    GetAllCards200Response:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/CardDTO'
        totalElements:
          type: integer
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean

    GetMyTransactions200Response:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/TransactionDTO'
        totalElements:
          type: integer
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean

    GetAllUsers200Response:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/UserDTO'
        totalElements:
          type: integer
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean

security:
  - bearerAuth: []

paths:
  /bankcards/api/v1/cards:
    post:
      summary: Create a new card
      tags: [Card Management]
      operationId: createCard
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCardRequest'
      responses:
        '201':
          description: Card created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardDTO'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden

  /bankcards/api/v1/cards/my-cards:
    get:
      summary: Get user's own cards
      tags: [Card Management]
      operationId: getMyCards
      security:
        - bearerAuth: []
      parameters:
        - name: maskedCardNumber
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/CardStatus'
        - name: minBalance
          in: query
          required: false
          schema:
            type: number
            format: double
        - name: maxBalance
          in: query
          required: false
          schema:
            type: number
            format: double
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: sort
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetMyCards200Response'
        '401':
          description: Unauthorized

  /bankcards/api/v1/cards/all:
    get:
      summary: Get all cards (Admin only)
      tags: [Card Management]
      operationId: getAllCards
      security:
        - bearerAuth: []
      parameters:
        - name: maskedCardNumber
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/CardStatus'
        - name: minBalance
          in: query
          required: false
          schema:
            type: number
            format: double
        - name: maxBalance
          in: query
          required: false
          schema:
            type: number
            format: double
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAllCards200Response'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden

  /bankcards/api/v1/cards/{cardId}:
    get:
      summary: Get card by ID
      tags: [Card Management]
      operationId: getCardById
      security:
        - bearerAuth: []
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardDTO'
        '404':
          description: Card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      summary: Update card status
      tags: [Card Management]
      operationId: updateCardStatus
      security:
        - bearerAuth: []
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCardStatusRequest'
      responses:
        '200':
          description: Status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardDTO'
        '400':
          description: Invalid status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete card (Admin only)
      tags: [Card Management]
      operationId: deleteCard
      security:
        - bearerAuth: []
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Card deleted successfully
        '404':
          description: Card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden

  /bankcards/api/v1/transfers/own-cards:
    post:
      summary: Transfer between own cards
      tags: [Transfer Management]
      operationId: transferBetweenOwnCards
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
      responses:
        '200':
          description: Transfer successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionDTO'
        '400':
          description: Invalid transfer request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - cards not owned by user

  /bankcards/api/v1/transfers/my-transactions:
    get:
      summary: Get user's transaction history
      tags: [Transfer Management]
      operationId: getMyTransactions
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetMyTransactions200Response'
        '401':
          description: Unauthorized

  /bankcards/api/v1/transfers/card/{cardId}/transactions:
    get:
      summary: Get transactions for a specific card
      tags: [Transfer Management]
      operationId: getCardTransactions
      security:
        - bearerAuth: []
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetMyTransactions200Response'
        '403':
          description: Access denied
        '404':
          description: Card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /bankcards/api/v1/admin/users:
    get:
      summary: Get all users (Admin only)
      tags: [Admin User Management]
      operationId: getAllUsers
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAllUsers200Response'
        '403':
          description: Forbidden

  /bankcards/api/v1/admin/users/{userId}:
    get:
      summary: Get user by ID
      tags: [Admin User Management]
      operationId: getUserById
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Update user
      tags: [Admin User Management]
      operationId: updateUser
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete user
      tags: [Admin User Management]
      operationId: deleteUser
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: User deleted successfully
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'